Move core types to shared
