+++ CopilotKit/packages/react-core/src/components/copilot-provider/copilot-provider.tsx
@@ -5,7 +5,8 @@ import { CopilotContext, CopilotApiConfig } from "../../context/copilot-context"
 import useTree from "../../hooks/use-tree";
 import { AnnotatedFunction } from "../../types/annotated-function";
 import { ChatCompletionCreateParams } from "openai/resources/chat";
-import { DocumentPointer, FunctionCallHandler } from "../../types";
+import { DocumentPointer } from "../../types";
+import { FunctionCallHandler } from "@copilotkit/shared";
 import useFlatCategoryStore from "../../hooks/use-flat-category-store";
 import { StandardCopilotApiConfig } from "./standard-copilot-api-config";
 import { CopilotProviderProps } from "./copilot-provider-props";
