+++ CopilotKit/packages/react-core/src/openai/chat-completion-client.ts
@@ -1,5 +1,5 @@
 import EventEmitter from "eventemitter3";
-import { Function, Message, Role } from "../types";
+import { Function, Message, Role } from "@copilotkit/shared";
 import {
   ChatCompletionTransport,
   ChatCompletionTransportFetchParams,
