--- CopilotKit/packages/react-core/src/types/base.ts
@@ -1,61 +1,0 @@
-export type Role = "system" | "user" | "assistant" | "function";
-
-export interface Message {
-  id: string;
-  createdAt?: Date;
-  content: string;
-  role: Role;
-  /**
-   * If the message has a role of `function`, the `name` field is the name of the function.
-   * Otherwise, the name field should not be set.
-   */
-  name?: string;
-  /**
-   * If the assistant role makes a function call, the `function_call` field
-   * contains the function call name and arguments. Otherwise, the field should
-   * not be set.
-   */
-  function_call?: FunctionCall;
-}
-
-export interface FunctionCall {
-  /**
-   * The arguments to call the function with, as generated by the model in JSON
-   * format. Note that the model does not always generate valid JSON, and may
-   * hallucinate parameters not defined by your function schema. Validate the
-   * arguments in your code before calling your function.
-   */
-  arguments?: string;
-  /**
-   * The name of the function to call.
-   */
-  name?: string;
-}
-
-export interface Function {
-  /**
-   * The name of the function to be called. Must be a-z, A-Z, 0-9, or contain
-   * underscores and dashes, with a maximum length of 64.
-   */
-  name: string;
-  /**
-   * The parameters the functions accepts, described as a JSON Schema object. See the
-   * [guide](/docs/guides/gpt/function-calling) for examples, and the
-   * [JSON Schema reference](https://json-schema.org/understanding-json-schema/) for
-   * documentation about the format.
-   *
-   * To describe a function that accepts no parameters, provide the value
-   * `{"type": "object", "properties": {}}`.
-   */
-  parameters: Record<string, unknown>;
-  /**
-   * A description of what the function does, used by the model to choose when and
-   * how to call the function.
-   */
-  description?: string;
-}
-
-export type FunctionCallHandler = (
-  chatMessages: Message[],
-  functionCall: FunctionCall,
-) => Promise<void>;
