+++ CopilotKit/packages/react-textarea/package.json
@@ -46,6 +46,7 @@
   },
   "dependencies": {
     "@copilotkit/react-core": "workspace:*",
+    "@copilotkit/shared": "workspace:*",
     "@emotion/css": "^11.11.2",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
