+++ CopilotKit/packages/react-ui/src/types/types.ts
@@ -1,4 +1,4 @@
-import { type Message } from "@copilotkit/react-core";
+import { type Message } from "@copilotkit/shared";
 
 export interface Chat extends Record<string, any> {
   id: string;
