+++ CopilotKit/packages/shared/src/types/openai-assistant.ts
@@ -1,3 +1,5 @@
+export type Role = "system" | "user" | "assistant" | "function";
+
 export interface FunctionCall {
   /**
    * The arguments to call the function with, as generated by the model in JSON
@@ -21,7 +23,7 @@ export interface Message {
   createdAt?: Date;
   content: string;
   ui?: string | null | undefined;
-  role: "system" | "user" | "assistant" | "function";
+  role: Role;
   /**
    * If the message has a role of `function`, the `name` field is the name of the function.
    * Otherwise, the name field should not be set.
@@ -34,6 +36,34 @@ export interface Message {
    */
   function_call?: string | FunctionCall;
 }
+
+export interface Function {
+  /**
+   * The name of the function to be called. Must be a-z, A-Z, 0-9, or contain
+   * underscores and dashes, with a maximum length of 64.
+   */
+  name: string;
+  /**
+   * The parameters the functions accepts, described as a JSON Schema object. See the
+   * [guide](/docs/guides/gpt/function-calling) for examples, and the
+   * [JSON Schema reference](https://json-schema.org/understanding-json-schema/) for
+   * documentation about the format.
+   *
+   * To describe a function that accepts no parameters, provide the value
+   * `{"type": "object", "properties": {}}`.
+   */
+  parameters: Record<string, unknown>;
+  /**
+   * A description of what the function does, used by the model to choose when and
+   * how to call the function.
+   */
+  description?: string;
+}
+
+export type FunctionCallHandler = (
+  chatMessages: Message[],
+  functionCall: FunctionCall,
+) => Promise<void>;
 
 export type AssistantMessage = {
   id: string;
